% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_incidence_ets_lfs_ons.R
\name{calculate_incidence_ets_lfs_ons}
\alias{calculate_incidence_ets_lfs_ons}
\title{Function to Calculate Incidence Rates from ETS and Demographic Data}
\usage{

  calculate_incidence_ets_lfs_ons(data_path = "~/data/tb_data/tbinenglanddataclean",
  ets_name = "clean_ets_2016.rds",
  demo_name = "E_ons_lfs_2000_2016.rds", return = TRUE, save = TRUE,
  incidence_name = "incidence",
  grouped_incidence_name = "age_grouped_incidence",
  condensed_grouped_incidence_name = "condensed_age_group_incidence",
  cases_demo_incidence_name = "cases_demo_incidence",
  save_path = "~/data/tb_data/tbinenglanddataclean",
  save_format = "rds", verbose = TRUE, theme_set = theme_minimal)
}
\arguments{
\item{data_path}{A charater string containing the file path to the demographic data.}

\item{ets_name}{A character string of the name for the cleaned ETS data.}

\item{demo_name}{A character string of the name for the cleanded
demographic ONS and LFS data.}

\item{return}{Logical, defaults to \code{TRUE}. Specifies whether to return cleanded data.}

\item{save}{Logical, defaults to \code{FALSE}. Specifies whether to save data}

\item{incidence_name}{A character string indicating the name to save the incidence data under.}

\item{grouped_incidence_name}{A character string indicating the name to save the age grouped
incidence data under.}

\item{condensed_grouped_incidence_name}{A character string indicating the name to save the condensed age
grouped data under.}

\item{cases_demo_incidence_name}{A character string indicating the name to save the cases, and demographic data under.}

\item{save_path}{The filepath for the data to be saved in}

\item{save_format}{A character vector specifying the format/formats to save the data into, defaults to rds. Currently
csv is also supported. See \code{\link[idmodelr]{save_data}} for details.}

\item{verbose}{A logical indicating whether summary information should be provided.}

\item{theme_set}{The ggplot theme to apply to the summary graphs, defaults to theme_minimal}
}
\value{
A list of dataframes containing incidence data in an increasingly condensed format.
}
\description{
This function returns multiple incidence rate datesets, with varying age
groupings. This enables more rapid interrogation of the data.
}
\examples{

}
