% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/case_rate.R
\name{case_rate}
\alias{case_rate}
\title{Function to Estimate the Case Rate for a Given Variable, stratified by given variables, dependant on age, year, agegrp2, and uk birth status. Missing data for the rate variable is dropped}
\usage{
case_rate(df, rate_for, strat_by = NULL, age_split = NULL,
  CoB_split = TRUE, Year_strat = TRUE)
}
\arguments{
\item{df}{A dataframe in the format of the ETS}

\item{rate_for}{A character string of the variable to calculate the rate for.}

\item{strat_by}{A character string of the variables to stratfy by.}

\item{age_split}{A character strig og the age split to stratify case counts by.}

\item{CoB_split}{A logical indicating if Country of birth should be used to split the case counts.}

\item{Year_strat}{A logical indicating if year should be stratified over.}
}
\value{
Returns a dataframe of case rates stratified by multiple variables with confidence intervals.
}
\description{
Function to estimate the case rate for a given variable, stratified by given variables,
dependant on age, year, agegrp2, and uk birth status. Missing data for the rate variable is dropped.
}
\examples{

}
