% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_munge_ETS_2016.R
\name{clean_munge_ets_2016}
\alias{clean_munge_ets_2016}
\title{Clean and Munge 2016 Enhanced Tuberculosis Surviellance Data}
\usage{
clean_munge_ets_2016(data_path = NULL, return = TRUE, save = TRUE,
  save_name = "clean_ets_2016", save_path = "~/data/tbinenglanddataclean",
  save_format = "rds", verbose = TRUE)
}
\arguments{
\item{data_path}{A character string containing the file pathway for the 2016 ETS data.}

\item{return}{Logical, defaults to \code{TRUE}. Specifies whether to return cleanded data.}

\item{save}{Logical, defaults to \code{FALSE}. Specifies whether to save data}

\item{save_name}{A character string containing the name to save the tidy ETS data under.}

\item{save_path}{A character string containing the file pathway to the folder into
which to save the tidy ETS data for the 2016 ETS data.}

\item{save_format}{A character vector specifying the format/formats to save the data into, defaults to rds. Currently
csv is also supported. See \code{\link[idmodelr]{save_data}} for details.}

\item{verbose}{A logical indicating whether summary information should be provided.}
}
\value{
A tidy tibble of TB notficiations in England from 2000 to 2016, with a row for
each notification.
}
\description{
A function that imports the ETS data as a stata file, converts it to rds
format, cleans it to a tidy format, and finally munges required variables for future analysis.
Data can be applied for \href{https://www.gov.uk/government/publications/tuberculosis-tb-in-england-surveillance-data}{here}
}
\details{
Function has 3 distinct sections:
\itemize{
\item Data Import
\item Setting factor variable baslines
\item Munge new variables
}
}
\examples{

}
