% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine_ons_with_lfs.R
\name{combine_ons_with_lfs}
\alias{combine_ons_with_lfs}
\title{A Function that Combines the ONS Demographic Data, with the LFS
Demographic Data}
\usage{
combine_ons_with_lfs(data_path = "~/data/tb_data/tbinenglanddataclean",
  ons_name = "E_demo_2000_2015.rds",
  lfs_name = "formatted_LFS_2000_2016.rds", countries = "England",
  return = TRUE, save = TRUE, save_name = "E_ons_lfs_2000_2016",
  save_path = "~/data/tb_data/tbinenglanddataclean", save_format = "rds",
  verbose = TRUE, interactive = TRUE, theme_set = theme_minimal)
}
\arguments{
\item{data_path}{A charater string containing the file path to the demographic data.}

\item{ons_name}{Character string of the file name of the ONS demographic data.}

\item{lfs_name}{Character string of the file name of the LFS demographic data.}

\item{countries}{A character string, the countries to include in the
dataset. By default only England is included. Note this is reliant on the data being present in the
demographic datasets.}

\item{return}{Logical, defaults to \code{TRUE}. Specifies whether to return cleanded data.}

\item{save}{Logical, defaults to \code{FALSE}. Specifies whether to save data}

\item{save_name}{A character string contaning the file name for the data to be saved under.}

\item{save_path}{The filepath for the data to be saved in}

\item{save_format}{A character vector specifying the format/formats to save the data into, defaults to rds. Currently
csv is also supported. See \code{\link[idmodelr]{save_data}} for details.}

\item{verbose}{A logical indicating whether summary information should be provided.}

\item{interactive}{Logical, defaults to \code{TRUE}. Should the summary graphs be returned as interactive objects.}

\item{theme_set}{The ggplot theme to apply to the summary graphs, defaults to theme_minimal}
}
\value{
A tidy tibble of demographic data by age between 2000 and 2015 for the specified countries
for both ONS and LFS data.
}
\description{
This function takes demographic data summarised by \code{\link[tbinenglanddataclean]{clean_demographics_uk}} and
\code{\link[tbinenglanddataclean]{clean_labour_force_survey}} and combines it into a single tidy dataset.
Summary statistics and plots can be returned to check both datasets.
}
\examples{

}
