% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/est_inc_crude_adj.R
\name{est_inc_crude_adj}
\alias{est_inc_crude_adj}
\title{Function to Estimate Age Adjusted Incidence}
\usage{
est_inc_crude_adj(filter, pop = pop_age, cases = cases_age,
  rate_pop = 1e+05)
}
\arguments{
\item{filter}{Country of Birth filter to use.}

\item{pop}{A dataframe of demographis.}

\item{cases}{A dataframe of cases.}

\item{rate_pop}{The incidence rate scale to use.}
}
\value{
A dataframe containing age stratified incidence.
}
\description{
Function to Estimate Age Adjusted Incidence
}
\examples{

}
