% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_case_counts.R
\name{extract_case_counts}
\alias{extract_case_counts}
\title{Extract Case Counts from the ETS Data}
\usage{
extract_case_counts(df, strat_var = c("age", "year", "ukborn"))
}
\arguments{
\item{df}{A dataframe in the current ETS format .}

\item{strat_var}{Vaiables to stratify case counts by.}
}
\value{
Case counts stratified by the required variables
}
\description{
Case counts stratified by the required variables. In addition this function adds additional grouping
variables for age so that the ETS data is inline with other demographic datasets.
}
\examples{

}
