% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/form_df_epiR_inc_rate.R
\name{form_df_epiR_inc_rate}
\alias{form_df_epiR_inc_rate}
\title{Function to Split by Country of Birth and Format for epiR}
\usage{
form_df_epiR_inc_rate(df, filter)
}
\arguments{
\item{df}{A dataframe.}

\item{filter}{Country of Birth filter to use.}
}
\value{
Returns a dataframe filtered by Country of Birth
}
\description{
Function to Split by Country of Birth and Format for epiR
}
\examples{

}
