% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_historic_tb_ew.R
\name{plot_historic_tb_ew}
\alias{plot_historic_tb_ew}
\title{Plot Historic TB Notifications in England and Wales}
\usage{
plot_historic_tb_ew(df = tb_not_ew, zoom_date = 1975,
  plot_theme = ggplot2::theme_minimal,
  colour_scale = ggplot2::scale_colour_viridis_d, return = FALSE)
}
\arguments{
\item{df}{A dataframe, (defaults to using \code{\link[tbinenglanddataclean]{tb_not_ew}})
of historic TB notifications}

\item{zoom_date}{Numeric, the year to filter notifications from for the second zoomed plot.}

\item{plot_theme}{The ggplot2 theme to use, defaults to \code{\link[ggplot2]{theme_minimal}}.}

\item{colour_scale}{The colour scale to plot with, defaults to \code{\link[ggplot2]{scale_colour_viridis_d}}}

\item{return}{A logical indiciting if the plot should be printed or returned (defaults to \code{FALSE})}
}
\value{
A ggplot2 plot of TB notifications over time, with a secondary zoomed plot from a specified date.
}
\description{
Plot Historic TB Notifications in England and Wales
}
\examples{
plot_historic_tb_ew()
}
