% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_pop_age_compare_ons_lfs.R
\name{plot_pop_age_compare_ons_lfs}
\alias{plot_pop_age_compare_ons_lfs}
\title{A Function to Plot Differences Between LFS and ONS Data}
\usage{
plot_pop_age_compare_ons_lfs(df)
}
\arguments{
\item{df}{A Dataframe as produced by \code{\link[tbinenglanddataclean]{combine_ons_with_lfs}} of
population demographics}
}
\value{
A ggplot summarising the differences between both ONS and LFS datasets
}
\description{
A Function to Plot Differences Between LFS and ONS Data
}
\examples{

}
