% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tb_not_ew.R
\docType{data}
\name{tb_not_ew}
\alias{tb_not_ew}
\title{Tuberculosis Notifications in England and Wales}
\format{A data frame with 102 rows and 4 variables.
\describe{
\item{year}{Calendar year}
\item{pulmonary}{The number of pulmonary cases (with or without extra-pulomnary disease)}
\item{extra_pulmonary}{The number of extra-pulmonary cases}
\item{total}{The total number of notified cases}
}}
\usage{
tb_not_ew
}
\description{
A dataset containing notified Tuberculosis cases from 1913 to 2014. This dataset was made available
by Public Health England and is available \href{https://www.gov.uk/government/uploads/system/uploads/attachment_data/file/464852/TB_case_notifications_1913_to_2014.pdf}{here}.
}
\keyword{datasets}
